\name{BayCausal-package}
\alias{BayCausal-package}
\alias{BayCausal}
\docType{package}
\title{
\packageTitle{BayCausal}
}
\description{
\packageDescription{BayCausal}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{BayCausal}

Maintainer: \packageMaintainer{BayCausal}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = BayCausal)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{BayCausal}}
%\section{Documentation Index}{\packageIndices{BayCausal}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
