% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_analysis.R
\name{extract_P_star}
\alias{extract_P_star}
\title{Extracts the posterior samples of P_star from the MCMC samples}
\usage{
extract_P_star(res)
}
\arguments{
\item{res}{A list containing MCMC samples}
}
\value{
posterior samples of P_star
}
\description{
Extracts the posterior samples of P_star from the MCMC samples
}
