% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_analysis.R
\name{extract_post_mean}
\alias{extract_post_mean}
\title{Extracts the posterior mean of a parameter from the MCMC samples}
\usage{
extract_post_mean(res, param, P_star)
}
\arguments{
\item{res}{A list containing MCMC samples}

\item{param}{The parameter to extract}

\item{P_star}{Only extract MCMC samples with this value of P_star}
}
\value{
posterior mean of the parameter
}
\description{
Extracts the posterior mean of a parameter from the MCMC samples for a given value of P_star
}
