% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_laplace.R
\name{fa_laplace}
\alias{fa_laplace}
\title{Title}
\usage{
fa_laplace(data, mh_setup_lst, init_lst, prior_lst, chain_setup_lst, verbose)
}
\arguments{
\item{data}{}

\item{mh_setup_lst}{}

\item{init_lst}{}

\item{prior_lst}{}

\item{chain_setup_lst}{}

\item{verbose}{}
}
\value{
a list containing mcmc samples
}
\description{
Title
}
