% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_toy_data_t.R
\name{generate_toy_data_t}
\alias{generate_toy_data_t}
\title{Generate t toy data}
\usage{
generate_toy_data_t(include_B = TRUE, include_X = TRUE, include_C = TRUE)
}
\arguments{
\item{include_B}{whether to include B in the toy data}

\item{include_X}{whether to include X in the toy data}

\item{include_C}{whether to include C in the toy data}
}
\value{
a list containing toy data ground truth
}
\description{
Generate t toy data for testing and example usage
}
