\name{glvcausal-package}
\alias{glvcausal-package}
\alias{glvcausal}
\docType{package}
\title{
\packageTitle{glvcausal}
}
\description{
\packageDescription{glvcausal}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{glvcausal}

Maintainer: \packageMaintainer{glvcausal}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = glvcausal)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{glvcausal}}
%\section{Documentation Index}{\packageIndices{glvcausal}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
