% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glvcausal_t.R
\name{glvcausal_t}
\alias{glvcausal_t}
\title{Implement the Bayesian causal discovery algorithm with Student-t errors}
\usage{
glvcausal_t(data, mh_setup_lst, init_lst, prior_lst, chain_setup_lst, verbose)
}
\arguments{
\item{data}{The observed data}

\item{mh_setup_lst}{A list containing the Metropolis-Hastings setups}

\item{init_lst}{A list containing the initial values of the parameters}

\item{prior_lst}{A list containing the prior hyper parameters. Must contain `nu_t` (degrees of freedom).}

\item{chain_setup_lst}{A list containing the MCMC chain setups}

\item{verbose}{A boolean indicating whether to print the progress of the MCMC chain}
}
\value{
a list containing mcmc samples
}
\description{
Implement the Bayesian causal discovery algorithm in the presence of cycles and unobserved confounders
assuming Student-t distributed errors
}
