% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingcausal_wo_x.R
\name{lingcausal_wo_x}
\alias{lingcausal_wo_x}
\title{Title}
\usage{
lingcausal_wo_x(
  data,
  mh_setup_lst,
  init_lst,
  prior_lst,
  chain_setup_lst,
  verbose
)
}
\arguments{
\item{data}{}

\item{mh_setup_lst}{}

\item{init_lst}{}

\item{prior_lst}{}

\item{chain_setup_lst}{}

\item{verbose}{}
}
\value{
a list containing mcmc samples
}
\description{
Title
}
