% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingfa_wo_x.R
\name{lingfa_wo_x}
\alias{lingfa_wo_x}
\title{Implement the Bayesian factor analysis algorithm}
\usage{
lingfa_wo_x(data, mh_setup_lst, init_lst, prior_lst, chain_setup_lst, verbose)
}
\arguments{
\item{data}{The observed data}

\item{mh_setup_lst}{A list containing the Metropolis-Hastings setups}

\item{init_lst}{A list containing the initial values of the parameters}

\item{prior_lst}{A list containing the prior hyper parameters}

\item{chain_setup_lst}{A list containing the MCMC chain setups}

\item{verbose}{A boolean indicating whether to print the progress of the MCMC chain}
}
\value{
a list containing mcmc samples
}
\description{
Implement the Bayesian factor analysis algorithm with independent laplace errors when no covariates are included
}
