% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_analysis.R
\name{post_process_L}
\alias{post_process_L}
\title{Title}
\usage{
post_process_L(res, order_flag, cutoff = Inf)
}
\arguments{
\item{res}{}

\item{order_flag}{}

\item{cutoff}{}
}
\value{
post processed samples
}
\description{
Title
}
