% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_default.R
\name{set_init_default}
\alias{set_init_default}
\title{Set default initial values for the chain}
\usage{
set_init_default(seed, P_star, data, prior_lst)
}
\arguments{
\item{seed}{random seed used to initialize the chain}

\item{P_star}{number of latent factors}

\item{data}{data list}

\item{prior_lst}{a list containing prior hyper parameters}
}
\value{
a list containing initial values
}
\description{
Other than the specified values, all the other values are generated randomly from their prior distributions. Tau is generated from full conditional distribution given all other quantities.
}
