% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_default.R
\name{set_prior_default_t}
\alias{set_prior_default_t}
\title{Set default prior hyper parameters for t-distribution model}
\usage{
set_prior_default_t(data, Eq = 1, nu_0 = 0.00025)
}
\arguments{
\item{data}{the data list}

\item{Eq}{expected number of non-zero elements in the sparsity matrix for each row}
}
\value{
a list containing chain setup
}
\description{
Set default prior hyper parameters for t-distribution model
}
