% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_B_t.R
\name{update_B_t}
\alias{update_B_t}
\title{Update B using Metropolis-Hastings with Student-t errors}
\usage{
update_B_t(
  B,
  gamma_beta,
  nu_beta,
  A,
  mu,
  L,
  C,
  sigma2,
  mh_step_size,
  data,
  nu_t
)
}
\arguments{
\item{B}{current B matrix}

\item{gamma_beta}{inclusion indicators}

\item{nu_beta}{variances}

\item{A}{coefficient matrix for X}

\item{mu}{mean vector}

\item{L}{factor loadings}

\item{C}{factor scores}

\item{sigma2}{noise variances (Normal scale parameter squared)}

\item{mh_step_size}{step size for MH}

\item{data}{data object}

\item{nu_t}{degrees of freedom for Student-t}
}
\description{
Update B using Metropolis-Hastings with Student-t errors
}
