% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_analysis.R
\name{extract_post_mean}
\alias{extract_post_mean}
\title{Title}
\usage{
extract_post_mean(res, param, P_star)
}
\arguments{
\item{res}{}

\item{param}{}

\item{P_star}{}
}
\value{
posterior mean of the parameter
}
\description{
Title
}
