% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_analysis.R
\name{extract_post_mean_all}
\alias{extract_post_mean_all}
\title{Title}
\usage{
extract_post_mean_all(res, param)
}
\arguments{
\item{res}{}

\item{param}{}
}
\value{
posterior mean of the parameter
}
\description{
Title
}
