% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_analysis.R
\name{post_process_L}
\alias{post_process_L}
\title{Title}
\usage{
post_process_L(res, order_flag, pos_sign, cutoff = -Inf)
}
\arguments{
\item{res}{A list containing MCMC samples}

\item{order_flag}{A logical indicating whether to order the columns of L}

\item{pos_sign}{A logical indicating whether to enforce positive sign on the first non-zero element of each column of L}

\item{cutoff}{A numeric value indicating the threshold below which the elements of L are set to zero}
}
\value{
post processed samples
}
\description{
Title
}
