% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_analysis.R
\name{post_process_L_mean}
\alias{post_process_L_mean}
\title{Title}
\usage{
post_process_L_mean(L_mean, order_flag, cutoff = Inf)
}
\arguments{
\item{L_mean}{}

\item{order_flag}{}

\item{cutoff}{}
}
\value{
post processed L posterior mean
}
\description{
Title
}
