% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_default.R
\name{set_init_default}
\alias{set_init_default}
\title{Title}
\usage{
set_init_default(seed, P_star, data, prior_lst)
}
\arguments{
\item{seed}{random seed}

\item{P_star}{number of latent factors}

\item{data}{data list}

\item{prior_lst}{a list containing prior hyperparameters}
}
\value{
a list containing initial values
}
\description{
Title
}
