% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_processed_init.R
\name{set_processed_init}
\alias{set_processed_init}
\title{Title}
\usage{
set_processed_init(
  data,
  mh_setup_lst,
  init_lst,
  prior_lst,
  chain_setup_lst_1,
  chain_setup_lst_2,
  verbose
)
}
\arguments{
\item{data}{}

\item{mh_setup_lst}{}

\item{init_lst}{}

\item{prior_lst}{}

\item{chain_setup_lst_1}{}

\item{chain_setup_lst_2}{}

\item{verbose}{}
}
\value{
processed initial values
}
\description{
Title
}
